// models/setting.ts
export interface ISetting {
  _id: string;
  about_me: string;
  intro_text: string;

  // personal info
  image: string;
  email: string;
  phone: string;
  address: string;

  // social media links
  github: string;
  linkedin: string;
  twitter: string;
  facebook: string;
  instagram: string;
  youtube: string;
}